livekit-agent/requirements.txt
livekit
livekit-agents
livekit-plugins-silero
python-dotenv
psutil>=7.0.0
importlib-metadata==7.2.1

